refactor: tests and 42 display
