+++ part1/source/tests.h
@@ -1,0 +1,10 @@
+#ifndef TESTS_H
+#define TESTS_H
+
+void print_42(void);
+void print_k_test();
+void print_test();
+void print_f_test();
+void main_tests();
+
+#endif
